v0.5.0 updated dotswebsocket to ver 2 removing all windows specific requirements allowing other OS builds added error logic for incompatible APWorld versions added better debug codes added the ability to press enter to send commands in console (NOTE: the text box must not be selected for this to work) added error logic in starting a game without any items or not correctly setting starting girl to be unlocked made it so item data is always printed to LogOutput.log

+++ ArchipelagoPackets/Packets.cs
@@ -169,8 +169,8 @@ namespace HuniePopArchiepelagoClient.ArchipelagoPackets
     }
     public class NetworkItem
     {
-        public int item;
-        public int location;
+        public long item;
+        public long location;
         public int player;
         public int flags;
     }
